Extra beveiliging toegevoegd. Controle of gebruiker wel degelijk admin is.

+++ edit/deleteLokaal.php
@@ -3,6 +3,10 @@
 include("../conf/db.php");
 include("../conf/sessionCheck.php");
 
+if(!$is_admin){
+	die("U heeft geen toegang tot deze pagina");
+}
+
 $lokaalID = mysql_real_escape_string($_GET['id']);
 
 $qry_delete = mysql_query("DELETE FROM lokalen WHERE id = '$lokaalID'");
